requirements.txt
gradio
fastapi
uvicorn
google-cloud-speech
google-generativeai
python-dotenv
